backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-multipart==0.0.12
numpy==2.1.2
opencv-python==4.10.0.84
torch==2.4.1
Pillow==10.4.0
